Expose mock client fixture to other endpoints' tests
